requirements.txt
Django==1.11.14
django-rest-framework==0.1.0
djangorestframework==3.8.2
MySQL-python==1.2.5
pkg-resources==0.0.0
pytz==2018.5
